flight-price-monitor/requirements.txt
anthropic>=0.40.0
amadeus>=8.0.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
click>=8.1.0
pydantic>=2.0.0
